+++ R/visualization.R
@@ -7,15 +7,17 @@
 #' @param input_index (numeric) a vector of length 1 or 2. 1 means the 1st input of the td, 2 means the second input of the td, and c(1, 2) means both of the inputs, which is the DEFAULT
 #' @param text_size (numeric) the size of plot ID, strip ID, and block ID numbers printed in the plots
 #' @param abline (logical) If TRUE, ab-lines are displayed as well. Default = FALSE. This applies only ton type = "rates" and type = "layout".
-#' @returns plot as a ggplot object
+#' @param leaflet (logical) If TRUE, the plot will be superimposed on a satellite imagery of the field. Default is FALSE. This option is effective only for type = "rates".
+#' @returns ggplot or leaflet (if leaflet == TRUE) object
 #' @import ggplot2
+#' @import leaflet
 #' @export
 #' @examples
 #' #--- load trial design ---#
 #' data(td_two_input)
 #' viz(td_two_input)
 #'
-viz <- function(td, type = "rates", input_index = c(1, 2), text_size = 3, abline = FALSE) {
+viz <- function(td, type = "rates", input_index = c(1, 2), text_size = 3, abline = FALSE, leaflet = FALSE) {
   #++++++++++++++++++++++++++++++++++++
   #+ Debug
   #++++++++++++++++++++++++++++++++++++
@@ -105,72 +107,179 @@ viz <- function(td, type = "rates", input_index = c(1, 2), text_size = 3, abline
     # tgt_rate_original <- gg_td$tgt_rate_original[[1]]
     # trial_design <- gg_td$trial_design[[1]]
 
-    gg_td <-
-      td_rows %>%
-      dplyr::mutate(data_for_plot = list(
-        data.table::data.table(rate = unique(trial_design$rate)) %>%
-          .[, rata_equiv := convert_rates(input_name, unit, gc_rate)] %>%
-          .[order(rate), ] %>%
-          .[, all_units := factor(paste(rate, " | ", rata_equiv))] %>%
-          dplyr::left_join(trial_design, ., by = "rate")
-      )) %>%
-      dplyr::mutate(need_equiv_rate = list(
-        data.table(data_for_plot)[, all(rate != rata_equiv)]
-      )) %>%
-      dplyr::mutate(legend_title = list(
-        get_legend_title(
-          unit_system,
-          need_equiv_rate,
-          input_name,
-          input_type,
-          unit
-        )
-      )) %>%
-      dplyr::mutate(g_tr = list(
-        ggplot() +
-          geom_sf(
-            data = field_sf,
-            fill = NA
-          ) +
-          geom_sf(
-            data = data_for_plot,
-            aes(fill = factor(all_units)),
-            color = "black"
-          ) +
-          scale_fill_brewer(name = legend_title, palette = "Greens") +
-          theme_void() +
-          ggtitle(
-            paste0(
-              "Trial design",
-              " (",
-              dplyr::case_when(
-                design_type == "ls" ~ "Latin Square",
-                design_type == "str" ~ "Strip",
-                design_type == "rstr" ~ "Randomized Strip",
-                design_type == "rb" ~ "Randomized Block",
-                design_type == "ejca" ~ "Extra Jump-conscious Alternate",
-                design_type == "sparse" ~ "Sparse"
-              ),
-              ")"
-            )
+    if (leaflet == TRUE) {
+      color_pal_1 <- colorFactor(palette = "Greens", domain = NULL)
+      color_pal_2 <- colorFactor(palette = "Oranges", domain = NULL)
+
+      data_for_leaflet <-
+        td_rows %>%
+        dplyr::mutate(data_for_plot = list(
+          data.table::data.table(rate = unique(trial_design$rate)) %>%
+            .[, rata_equiv := convert_rates(input_name, unit, rate)] %>%
+            .[order(rate), ] %>%
+            .[, all_units := factor(paste(rate, " | ", rata_equiv))] %>%
+            .[rata_equiv == rate, all_units := factor(rate)] %>%
+            dplyr::left_join(trial_design, ., by = "rate")
+        )) %>%
+        dplyr::mutate(need_equiv_rate = list(
+          data.table(data_for_plot)[, all(rate != rata_equiv)]
+        )) %>%
+        dplyr::mutate(legend_title = list(
+          get_legend_title(
+            unit_system,
+            need_equiv_rate,
+            input_name,
+            input_type,
+            unit
           )
-      )) %>%
-      dplyr::mutate(g_fig = list(
-        if (abline == TRUE) {
-          g_tr +
-            geom_sf(data = ab_lines, aes(color = "applicator/planter ab-line")) +
-            geom_sf(data = harvest_ab_lines, aes(color = "harvester ab-line")) +
-            scale_color_manual(
-              name = "",
-              values = c(
-                "applicator/planter ab-line" = "red", "harvester ab-line" = "blue"
+        ))
+
+      legend_titles <- data_for_leaflet$legend_title %>% unlist()
+      layer_titles <-
+        data_for_leaflet$input_name %>%
+        lapply(to_title) %>%
+        unlist()
+
+      if (nrow(td_rows) == 2) {
+        leaflet_map <-
+          leaflet() %>%
+          addPolygons(
+            data = data_for_leaflet$data_for_plot[[1]],
+            fillOpacity = 0.6,
+            color = "black",
+            weight = 1,
+            fillColor = ~ color_pal_1(all_units),
+            group = layer_titles[1]
+          ) %>%
+          addPolygons(
+            data = data_for_leaflet$data_for_plot[[2]],
+            fillOpacity = 0.6,
+            color = "black",
+            weight = 1,
+            fillColor = ~ color_pal_2(all_units),
+            group = layer_titles[2]
+          ) %>%
+          addProviderTiles(
+            providers$Esri.WorldImagery
+          ) %>%
+          addLegend(
+            data = data_for_leaflet$data_for_plot[[1]],
+            pal = color_pal_1,
+            position = "topleft",
+            values = ~all_units,
+            title = legend_titles[[1]],
+            group = layer_titles[1]
+          ) %>%
+          addLegend(
+            data = data_for_leaflet$data_for_plot[[2]],
+            pal = color_pal_2,
+            position = "topleft",
+            values = ~all_units,
+            title = legend_titles[[2]],
+            group = layer_titles[2]
+          ) %>%
+          addLayersControl(
+            overlayGroups = c(layer_titles[1], layer_titles[2]),
+            options = layersControlOptions(collapsed = FALSE)
+          ) %>%
+          hideGroup(layer_titles[2])
+      } else {
+        leaflet_map <-
+          leaflet() %>%
+          addPolygons(
+            data = data_for_leaflet$data_for_plot[[1]],
+            fillOpacity = 0.6,
+            color = "black",
+            weight = 1,
+            fillColor = ~ color_pal_1(all_units),
+            group = layer_titles[1]
+          ) %>%
+          addProviderTiles(
+            providers$Esri.WorldImagery
+          ) %>%
+          addLegend(
+            data = data_for_leaflet$data_for_plot[[1]],
+            pal = color_pal_1,
+            position = "topleft",
+            values = ~all_units,
+            title = legend_titles[[1]],
+            group = layer_titles[1]
+          ) %>%
+          addLayersControl(
+            overlayGroups = c(layer_titles[1]),
+            options = layersControlOptions(collapsed = FALSE)
+          )
+      }
+    } else {
+
+      gg_td <-
+        td_rows %>%
+        dplyr::mutate(data_for_plot = list(
+          data.table::data.table(rate = unique(trial_design$rate)) %>%
+            .[, rata_equiv := convert_rates(input_name, unit, rate)] %>%
+            .[order(rate), ] %>%
+            .[, all_units := factor(paste(rate, " | ", rata_equiv))] %>%
+            .[rata_equiv == rate, all_units := factor(rate)] %>%
+            dplyr::left_join(trial_design, ., by = "rate")
+        )) %>%
+        dplyr::mutate(need_equiv_rate = list(
+          data.table(data_for_plot)[, all(rate != rata_equiv)]
+        )) %>%
+        dplyr::mutate(legend_title = list(
+          get_legend_title(
+            unit_system,
+            need_equiv_rate,
+            input_name,
+            input_type,
+            unit
+          )
+        )) %>%
+        dplyr::mutate(g_tr = list(
+          ggplot() +
+            geom_sf(
+              data = field_sf,
+              fill = NA
+            ) +
+            geom_sf(
+              data = data_for_plot,
+              aes(fill = factor(all_units)),
+              color = "black"
+            ) +
+            scale_fill_brewer(name = legend_title, palette = "Greens") +
+            theme_void() +
+            ggtitle(
+              paste0(
+                "Trial design",
+                " (",
+                dplyr::case_when(
+                  design_type == "ls" ~ "Latin Square",
+                  design_type == "str" ~ "Strip",
+                  design_type == "rstr" ~ "Randomized Strip",
+                  design_type == "rb" ~ "Randomized Block",
+                  design_type == "ejca" ~ "Extra Jump-conscious Alternate",
+                  design_type == "sparse" ~ "Sparse"
+                ),
+                ")"
               )
             )
-        } else {
-          g_tr
-        }
-      )) %>%
-      dplyr::select(g_fig)
+        )) %>%
+        dplyr::mutate(g_fig = list(
+          if (abline == TRUE) {
+            g_tr +
+              geom_sf(data = ab_lines, aes(color = "applicator/planter ab-line")) +
+              geom_sf(data = harvest_ab_lines, aes(color = "harvester ab-line")) +
+              scale_color_manual(
+                name = "",
+                values = c(
+                  "applicator/planter ab-line" = "red", "harvester ab-line" = "blue"
+                )
+              )
+          } else {
+            g_tr
+          }
+        )) %>%
+        dplyr::select(g_fig)
+    }
   } else if (type == "layout") {
     gg_td <-
       td_rows %>%
@@ -234,8 +343,9 @@ viz <- function(td, type = "rates", input_index = c(1, 2), text_size = 3, abline
     stop("The type you specified is not one of the allowed options.")
   }
 
-
-  if (nrow(gg_td) > 1) {
+  if (leaflet == TRUE & type == "rates") {
+    leaflet_map
+  } else if (nrow(gg_td) > 1) {
     ggpubr::ggarrange(gg_td$g_fig[[1]], gg_td$g_fig[[2]], ncol = 2)
   } else {
     gg_td$g_fig[[1]]