+++ NEWS.md
@@ -1,4 +1,10 @@
-# ofpetrial 0.1.0
+# ofpetrial 0.1.1
+
++ option of using leaflet to see experimental plots on a satellite image of the field was added to `viz()` 
++ `create_strips()` (an internal function) was overhauled to pass R CMD check on Fedora with ATLAS. Specifically, `st_tilt()`, which used to be used in the function, has been abandoned.
++ fixed the bug in `viz()` which presented `gc_rate`
+
+# ofpetrial 0.1
 
 The first complete set of functions that make the package functional and useful for the users to design and implement on-farm precision experiments.